Use the set name in the JUnit reports when avalible
We were running into an issue where Bamboo was treating each test as different from last time it was ran because it had a different (uniquely generated) class name based on the test_set. This meant that a test suite that has 3 failures would show as 3 *new* failures each time it's run rather than 3 *existing* failures.